Fixes to gcc warnings on v6.3.0

+++ include/string.h
@@ -8,7 +8,7 @@ typedef unsigned int size_t;
 void *memcpy(void *dst_ptr, const void *src_ptr, size_t count);
 void *memsetw(void *dst, int val, size_t count);
 void *memsetd(void *dst, int val, size_t count);
-size_t strlen(const char *str);
+unsigned strlen(const char *str);
 int strcmp(const char *s1, const char *s2);
 
 #define offsetof(str, mbr) (&(str.mbr) - &str)
